refactored event-map accordingly to prepare for future updates
